+++ mvc/controllers/homepage.js
@@ -1,8 +1,9 @@
 module.exports = (router, app) => {
-  router.route('/').get((req, res) => {
-    let model = require('models/global')(req, res)
-    model = require('models/homepage')(model)
-    model.content.pageTitle = 'Homepage'
-    res.render('homepage', model)
-  })
-}
+  router.route("/").get((req, res) => {
+    let model = require("models/global")(req, res);
+    model.routePrefix = app.get("routePrefix") || "";
+    model = require("models/homepage")(model);
+    model.content.pageTitle = "Homepage";
+    res.render("homepage", model);
+  });
+};