+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## Upcoming Changes
 
+- feat: Route prefix support. You can run this app at a path prefix by setting the `PKMN_TCG_APP_ROUTE_PREFIX` environment variable
 - feat: Updated cronjobs to use system timezone rather than just EST
 
 ## 1.0.0 - (Oct 19, 2024)